pyproject.toml
[tool.poetry]
name = "cash-tg-bot"
version = "0.1.0"
description = ""
authors = ["vgerasimov <vgerasimov@quantumbrains.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.0"
pytelegrambotapi = "^4.15.3"
loguru = "^0.7.2"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
